refactor(niri): move the IR rewrites into their own module
The projection says what an expression denotes; `rewrite` is what becomes of the
expression once it does. It reaches the IR only through `EditSink`, which is
what lets the same rewrites serve the throwaway body a frame runs on and the
real one an engine keeps coherent.

That completes the split. `niri` keeps the interpreter, the lattice and the maps
the engine is configured with; each of the five modules under it answers one
question, and the `&self` / `&mut self` line between reading an expression and
running one is now a module boundary rather than a convention.
